+++ dynamic-mapping-ui/src/mapping/step-one/mapping-properties.component.html
@@ -21,7 +21,7 @@
 
 <div class="p-16 p-t-0 flex-no-shrink separator-bottom">
   <div class="row">
-    <div class="col-lg-4 col-lg-offset-4">
+    <div class="col-lg-6 col-lg-offset-3">
       <h4 class="text-medium">
         <div class="text-center">
           Define properties for mapping: {{ mapping.mappingType }}&nbsp;
